Software/releases/smartdesk3/3.0.0/system/requirements-server.txt
simplejson
requests==2.11.1
enum
gtts
pyzmq
#spotify
pyowm
#urllib2
#urllib3
#apiclient
discovery
#oauth2client
python-dateutil
pytz
python-twitter
uber_rides
mysql-connector
netifaces
mock
